Extract and rename OakScheduler class
Fixes: #7
